Move classes to shared code.
